+++ pkg/chaosctl/recover/recover.go
@@ -17,6 +17,10 @@ package recover
 
 import (
 	"context"
+	"fmt"
+	"strings"
+
+	"github.com/pkg/errors"
 
 	ctrlclient "github.com/chaos-mesh/chaos-mesh/pkg/ctrl/client"
 )
@@ -33,19 +37,46 @@ type PartialPod struct {
 	Iptables []string
 }
 
-type Recover interface {
+type Recoverer interface {
 	// Recover target pod forcedly
 	Recover(ctx context.Context, pod *PartialPod) error
 }
 
-type RecoverBuilder func(client *ctrlclient.CtrlClient) Recover
+type RecovererBuilder func(client *ctrlclient.CtrlClient) Recoverer
 
-type noopRecover struct{}
-
-func NewNoopRecover(client *ctrlclient.CtrlClient) Recover {
-	return &noopRecover{}
+type cleanProcessRecoverer struct {
+	client  *ctrlclient.CtrlClient
+	process string
 }
 
-func (r *noopRecover) Recover(ctx context.Context, pod *PartialPod) error {
+func newCleanProcessRecoverer(client *ctrlclient.CtrlClient, process string) Recoverer {
+	return &cleanProcessRecoverer{
+		client:  client,
+		process: process,
+	}
+}
+
+func (r *cleanProcessRecoverer) Recover(ctx context.Context, pod *PartialPod) error {
+	var pids []string
+	for _, process := range pod.Processes {
+		if process.Command == r.process {
+			pids = append(pids, process.Pid)
+		}
+	}
+
+	if len(pids) == 0 {
+		printStep(fmt.Sprintf("all %s processes are cleaned up", r.process))
+		return nil
+	}
+	printStep(fmt.Sprintf("cleaning %s processes: %v", r.process, pids))
+
+	killedPids, err := r.client.KillProcesses(ctx, pod.Namespace, pod.Name, pids)
+	if err != nil {
+		return errors.Wrapf(err, "kill %s processes", r.process)
+	}
+	if len(killedPids) != 0 {
+		printStep(fmt.Sprintf("%s processes(%s) are cleaned up", r.process, strings.Join(killedPids, ", ")))
+	}
+
 	return nil
 }